FEAT-#1831: Move TracingWrappingConnection to separate package
